Project/Data-Engineering/Duckdb/requirements.txt
polars
pandas
numpy
apache-airflow==2.8.3
psycopg2-binary
duckdb
SQLAlchemy